+++ Programmer.h
@@ -1,0 +1,13 @@
+#ifndef PROGRAMMER_H
+#define PROGRAMMER_H
+
+using namespace std;
+
+struct Programmer {
+    string Name;
+    string Gender;
+    string BirthYear;
+    string DeadYear;
+};
+
+#endif
